Suppress compiler warning

+++ Modules/socketmodule.c
@@ -1902,7 +1902,7 @@ getsockaddrarg(PySocketSockObject *s, PyObject *args,
             addr->sun_path[path.len] = 0;
 
             /* including the tailing NUL */
-            *len_ret = path.len + offsetof(struct sockaddr_un, sun_path) + 1;
+            *len_ret = (int)path.len + offsetof(struct sockaddr_un, sun_path) + 1;
         }
         addr->sun_family = s->sock_family;
         memcpy(addr->sun_path, path.buf, path.len);
